+++ Rover/mode_smart_rtl.cpp
@@ -1,5 +1,8 @@
 #include "Rover.h"
 
+
+#if MODE_SMARTRTL_ENABLED
+
 bool ModeSmartRTL::_enter()
 {
     // SmartRTL requires EKF (not DCM)
@@ -135,3 +138,6 @@ void ModeSmartRTL::save_position()
     const bool save_pos = (rover.control_mode != &rover.mode_smartrtl);
     g2.smart_rtl.update(true, save_pos);
 }
+
+
+#endif  // MODE_SMARTRTL_ENABLED